+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Updated [status colours](/#/Theme?id=colors): danger, success, warning.
 - Add `fill` and `stroke` props on [Icon](/#/Components?id=icon).
 - Display all icons in the [Wardrobe Styleguide UI](/#/Components?id=icon).
+- Allow [Charts](/#/Charts) updates.
 
 ## [3.10.2] - 2019-03-26